Move last utilities to shared.
