+++ components/server/src/external/routes/measurement.py
@@ -8,10 +8,10 @@ from typing import cast
 import bottle
 from pymongo.database import Database
 
-from server_utilities.type import MetricId, SourceId
 from shared.database.measurements import insert_new_measurement, latest_measurement
 from shared.model.measurement import Measurement
 from shared.model.metric import Metric
+from shared.utils.type import MetricId, SourceId
 
 from ..database import sessions
 from ..database.datamodels import latest_datamodel
