+++ components/server/src/external/routes/notification.py
@@ -3,7 +3,7 @@
 import bottle
 from pymongo.database import Database
 
-from server_utilities.type import ReportId
+from shared.utils.type import ReportId
 
 from ..database.datamodels import latest_datamodel
 from ..database.reports import insert_new_report, latest_reports
