+++ components/server/src/external/utils/functions.py
@@ -21,7 +21,7 @@ from cryptography.fernet import Fernet
 from lxml.html.clean import autolink_html, clean_html  # noqa: DUO107, # nosec, pylint: disable=no-name-in-module
 
 from shared.utils.functions import iso_timestamp
-from server_utilities.type import ReportId
+from shared.utils.type import ReportId
 
 from .type import URL
